+++ Mall/src/main/java/com/example/mall/dto/user/UserEditBalanceDto.java
@@ -3,6 +3,8 @@ package com.example.mall.dto.user;
 import io.swagger.annotations.ApiParam;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 用户编辑价格 dto
  * @author By-Lin
@@ -10,8 +12,8 @@ import lombok.Data;
 @Data
 public class UserEditBalanceDto {
     @ApiParam(required = true)
-    private int id;
+    private Integer id;
 
     @ApiParam(required = true)
-    private double balance;
+    private BigDecimal balance;
 }
