+++ Mall/src/main/java/com/example/mall/dto/user/UserEditDto.java
@@ -3,6 +3,8 @@ package com.example.mall.dto.user;
 import io.swagger.annotations.ApiParam;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 用户修改 dto
  * @author By-Lin
@@ -10,15 +12,15 @@ import lombok.Data;
 @Data
 public class UserEditDto {
     @ApiParam(required = true)
-    private int id;
+    private Integer id;
 
     private String avatar;
     private String userName;
     private String phone;
 
     @ApiParam(required = true)
-    private double balance;
+    private BigDecimal balance;
 
     @ApiParam(required = true)
-    private int status;
+    private Integer status;
 }
