+++ Mall/src/main/java/com/example/mall/entity/Goods.java
@@ -2,6 +2,7 @@ package com.example.mall.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -10,12 +11,12 @@ import java.util.Date;
  */
 @Data
 public class Goods {
-    private int goodsId;
+    private Integer goodsId;
     private String goodsName;
-    private double goodsPrice;
-    private int goodsStatus;
-    private int storeCateId;
-    private int goodsCount;
-    private int isDel;
+    private BigDecimal goodsPrice;
+    private Integer goodsStatus;
+    private Integer storeCateId;
+    private Integer goodsCount;
+    private Integer isDel;
     private Date createTime;
 }
