+++ Mall/src/main/java/com/example/mall/entity/User.java
@@ -2,6 +2,7 @@ package com.example.mall.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -10,12 +11,12 @@ import java.util.Date;
  */
 @Data
 public class User {
-    private int id;
+    private Integer id;
     private String avatar;
     private String userName;
     private String password;
     private String phone;
-    private double balance;
-    private int status;
+    private BigDecimal balance;
+    private Integer status;
     private Date registTime;
 }
